Removing BCP38 section
After careful review, this is far too complex a topic to deal with in
this document. There is a lot more value if this was handled in a
separate push than peering -- because most of this work deals with
the access layer rather than BGP peering. We certainly don't want to be
telling people to do something silly like turning on uRPF checks because
they are going to be burned if someone advertises more specifics which
are received by transit and not via a peer, then get blocked due to RPF
checks.